requirements.txt
streamlit
pandas
numpy
plotly
pywebviewgit init
